imp: organize message handling methods under handlers module
