--- clients/sov-celestia-cw/src/contexts/mod.rs
@@ -1,6 +1,0 @@
-pub mod client_ctx;
-pub mod core_ctx;
-pub mod custom_ctx;
-mod definition;
-
-pub use definition::{Context, StorageMut, StorageRef};
